Added styles for payment form in CO
! IMPORTANT ! There is a serious bug in the code, that prevents
checkbox and radio buttons from being checked in about 50% times,
especially, if clicked on on the label and not on the input itself.
This is a current iOS & Chrome bug and it requires a major refactoring
of code and replacement of labels with something else, that can
register taps surely. It is also possible, that the whole concept must
be changed.